Add Tree of Life
This was our final project for CS162. This text based game requires you to travel to several worlds and interact with characters in order to collect powerful enough weapons to defeat a final boss. The requirements were to utilize inheritance and polymorphism, as well as linked lists, as I have done through the inherited Character classes and through connecting the worlds with pointers to base class World. It manually manages dynamically allocated memory and is free of memory leaks.